api/requirements.txt
flask
Flask-SQLAlchemy
flask-restful
pyyaml
pandas
numpy